# docs/api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Workspace Intelligence API
  description: |
    REST API for the Workspace Intelligence System - AI-powered code search, understanding, and completion.

    ## Features
    - Semantic code search across workspaces
    - AI-powered code completion
    - Context-aware code understanding
    - Real-time workspace indexing

    ## Authentication
    All endpoints require JWT authentication via `Authorization: Bearer <token>` header.

  version: 1.0.0
  contact:
    name: API Support
    email: api@workspaceintelligence.com
  license:
    name: Proprietary
servers:
  - url: https://api.workspaceintelligence.com/api/v1
    description: Production server
  - url: https://staging-api.workspaceintelligence.com/api/v1
    description: Staging server
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - BearerAuth: []
...
paths:
...
  /search/semantic:
    post:
      tags:
        - Search
      summary: Semantic code search
      description: Search for code using natural language or code snippets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              naturalLanguage:
                summary: Natural language query
                value:
                  workspace_id: "550e8400-e29b-41d4-a716-446655440000"
                  query: "function that handles user authentication"
                  limit: 10
              codeSnippet:
                summary: Code snippet query
                value:
                  workspace_id: "550e8400-e29b-41d4-a716-446655440000"
                  query: "public boolean authenticate(String username, String password)"
                  filters:
                    language: "java"
                    element_type: "method"
                  limit: 5
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SearchRequest:
      type: object
      required:
        - workspace_id
        - query
      properties:
        workspace_id:
          type: string
          format: uuid
        query:
          type: string
          minLength: 1
          maxLength: 1000
        filters:
          type: object
          properties:
            language:
              type: array
              items:
                type: string
                enum: [java, python, javascript, typescript, go, rust, cpp]
            element_type:
              type: array
              items:
                type: string
                enum: [function, method, class, interface, variable]
            file_pattern:
              type: string
              example: "src/**/*.java"
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 10
        offset:
          type: integer
          minimum: 0
          default: 0
    SearchResponse:
      type: object
      properties:
        query_id:
          type: string
          format: uuid
        results:
          type: array
          items:
            $ref: '#/components/schemas/CodeElement'
        total:
          type: integer
        latency_ms:
          type: integer
    CodeElement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        score:
          type: number
          format: float
          minimum: 0
          maximum: 1
        element_type:
          type: string
          enum: [function, method, class, interface, variable]
        name:
          type: string
        qualified_name:
          type: string
        file_path:
          type: string
        start_line:
          type: integer
        end_line:
          type: integer
        language:
          type: string
        signature:
          type: string
        docstring:
          type: string
        content:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "BAD_REQUEST"
              message: "Invalid request parameters"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              code: "UNAUTHORIZED"
              message: "Invalid or missing authentication token"
...